requirements.txt
discord.py>=2.3.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
sentencepiece>=0.1.99
python-dotenv>=1.0.0
aiosqlite>=0.19.0
numpy>=1.24.0
scipy>=1.10.0
psutil>=5.9.0
protobuf>=3.20.0